ft: move workspace homepage to `/`
