+++ packages/hash/frontend/src/pages/signup.page/signup-intro.tsx
@@ -39,7 +39,7 @@ export const SignupIntro: FunctionComponent<SignupIntroProps> = ({
 
   useEffect(() => {
     if (authenticatedUser?.accountSignupComplete) {
-      void router.push(`/${authenticatedUser.userAccountId}`);
+      void router.push("/");
     }
   }, [authenticatedUser, router]);
